requirements.txt
streamlit
streamlit_option_menu
streamlit_lottie
Pillow
requests